Policy: file-scope clamp helper, split filter empty row into own file
cmux-policy findings: pure helpers should be file-scope private funcs
rather than private statics, and new files should hold one major type.
Moves clampedWorkspacePreviewLineCount to file scope (@MainActor, since
the range constant lives on the @MainActor settings class) and gives
WorkspaceListFilterEmptyRow its own file.
